+++ packages/article-image/__tests__/ios/jest.config.js
@@ -1,9 +1,3 @@
-const jestConfigurator = require("@times-components/jest-configurator");
+const jestConfigurator = require("@times-components/jest-configurator").default;
 
-module.exports = Object.assign(jestConfigurator("article-image", "ios"), {
-  collectCoverageFrom: [
-    "**/packages/article-image/*.js",
-    "!**/packages/article-image/*stories.js",
-    "!**/packages/article-image/*.web.js"
-  ]
-});
+module.exports = jestConfigurator("article-image", "ios");