+++ packages/article-label/__tests__/android/jest.config.js
@@ -1,9 +1,3 @@
-const jestConfigurator = require("@times-components/jest-configurator");
+const jestConfigurator = require("@times-components/jest-configurator").default;
 
-module.exports = Object.assign(jestConfigurator("article-label", "android"), {
-  collectCoverageFrom: [
-    "**/packages/article-label/article-label.js",
-    "**/packages/article-label/style/shared.js",
-    "**/packages/article-label/**/*.android.js"
-  ]
-});
+module.exports = jestConfigurator("article-label", "android");