+++ packages/article-label/__tests__/web/jest.config.js
@@ -1,9 +1,3 @@
-const jestConfigurator = require("@times-components/jest-configurator");
+const jestConfigurator = require("@times-components/jest-configurator").default;
 
-module.exports = Object.assign(jestConfigurator("article-label", "web"), {
-  collectCoverageFrom: [
-    "**/packages/article-label/article-label.js",
-    "**/packages/article-label/style/*.js",
-    "!**/packages/article-label/**/*.android.js"
-  ]
-});
+module.exports = jestConfigurator("article-label", "web");