+++ packages/author-profile/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.37.10"></a>
+## [0.37.10](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.37.9...@times-components/author-profile@0.37.10) (2018-01-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/author-profile
+
 <a name="0.37.9"></a>
 ## [0.37.9](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.37.8...@times-components/author-profile@0.37.9) (2018-01-05)