+++ packages/provider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.20.5"></a>
+## [0.20.5](https://github.com/newsuk/times-components/compare/@times-components/provider@0.20.4...@times-components/provider@0.20.5) (2018-01-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/provider
+
 <a name="0.20.4"></a>
 ## [0.20.4](https://github.com/newsuk/times-components/compare/@times-components/provider@0.20.3...@times-components/provider@0.20.4) (2018-01-05)