+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.0.11"></a>
+## [0.0.11](https://github.com/newsuk/times-components/compare/@times-components/utils@0.0.10...@times-components/utils@0.0.11) (2018-01-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/utils
+
 <a name="0.0.10"></a>
 ## [0.0.10](https://github.com/newsuk/times-components/compare/@times-components/utils@0.0.9...@times-components/utils@0.0.10) (2018-01-05)